LOADER
- dodano klasę loadera
- dodano fcję wczytujcą modele
- dodano dużo bibliotek do maina
- przeniesiono ogólne fcje z main.h do main_file.h
- blok ma teraz współrzędne we float a nie w int

+++ Arkanoid3D/Block.h
@@ -5,14 +5,16 @@
 
 class Block {
 private:
-	int X;	//pozycja X bloczka
-	int Y;	//pozycja Y bloczka
+	float X;	//pozycja X bloczka
+	float Y;	//pozycja Y bloczka
 	int health;	//ile uderzeñ zosta³o do zniszczenia bloczka
 	bool destroyed;	//czy zosta³ zniszczony
+	
+
 	//TODO model i tekstura
 public:
 	//TODO konstruktor i destruktor
-	Block(int x, int y) {
+	Block(float x, float y) {
 		this->X = x;
 		this->Y = y;
 		this->health = 1;
